+++ src/ccSelectBeadTypeInCylinder.cpp
@@ -284,7 +284,7 @@ bool ccSelectBeadTypeInCylinder::Pack(const tguArgumentSequence& vArguments)
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(ccSelectBeadTypeInRegion::Pack(vArguments) && vArguments.size() == GetArgumentTotal())
+    if(ccSelectBeadTypeInRegion::Pack(vArguments) && static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         vArguments.at(2)->GetValue(&m_X);
         vArguments.at(3)->GetValue(&m_Y);
