+++ src/ccSelectBeadTypeInSlice.cpp
@@ -280,7 +280,7 @@ bool ccSelectBeadTypeInSlice::Pack(const tguArgumentSequence& vArguments)
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(ccSelectBeadTypeInRegion::Pack(vArguments) && vArguments.size() == GetArgumentTotal())
+    if(ccSelectBeadTypeInRegion::Pack(vArguments) && static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         vArguments.at(2)->GetValue(&m_X);
         vArguments.at(3)->GetValue(&m_Y);
