+++ src/ccSelectPolymerTypeHeadInEllipsoid.cpp
@@ -300,7 +300,7 @@ bool ccSelectPolymerTypeHeadInEllipsoid::Pack(const tguArgumentSequence& vArgume
 {
 #if EnableCommandGroups == SimCommandEnabled
 
-    if(ccSelectPolymerTypeInRegion::Pack(vArguments) && vArguments.size() == GetArgumentTotal())
+    if(ccSelectPolymerTypeInRegion::Pack(vArguments) && static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         vArguments.at(2)->GetValue(&m_XC);
         vArguments.at(3)->GetValue(&m_YC);
