+++ src/ctAssignExistingBeadType.cpp
@@ -191,7 +191,7 @@ bool ctAssignExistingBeadType::IsDataValid(const CInputData& riData) const
 
 bool ctAssignExistingBeadType::Pack(const tguArgumentSequence& vArguments)
 {
-    if(vArguments.size() == GetArgumentTotal())
+    if(static_cast<long>(vArguments.size()) == GetArgumentTotal())
     {
         ctApplyCommand::Pack(vArguments);
         vArguments.at(1)->GetValue(&m_AssignedName);
